\name{snowcat-package}
\alias{snowcat-package}
\alias{snowcat}
\docType{package}
\title{
    SNOWCAT: Streamlining Our Workflow: Code Alternative for TRACTOR
}
\description{

    SNOWCAT: \strong{S}treamli\strong{n}ing \strong{O}ur \strong{W}orkflow: \strong{C}ode \strong{A}lternative for \strong{T}RACTOR

    Snowcat is a Genome-wide association study (GWAS) pipeline with
    correction for local ancestry as estimated with RFMIX.
    It follows the model originally introduced by Elizabeth G. Atkinson in the TRACTOR pipeline
    (see \url{https://www.nature.com/articles/s41588-020-00766-y}, \url{https://github.com/Atkinson-Lab/Tractor}).

}
\details{
    \tabular{ll}{
        Package: \tab snowcat\cr
        Type: \tab Package\cr
        License: \tab LGPL-3 \cr
        LazyLoad: \tab yes\cr
        Depends: \tab methods\cr
    }
}
\author{
    Andrey A Shabalin \email{andrey.shabalin@gmail.com}

    Maintainer: Andrey A Shabalin <andrey.shabalin@gmail.com>
}
\keyword{ package }
\keyword{ snowcat }
